LandDetector: switch to .hpp convention instead of .h
